[TASK] Code Improvements
